requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
sqlite3==3.35.0
flask-login==0.6.3
werkzeug==3.0.1 